Don't download image if it exists already

+++ check_gnocchi_service.py
@@ -78,8 +78,13 @@ def create_new_resource(resource_type, resource_name=""):
         if resource_name == "":
             resource_name = "centos-7-image"
         # get image
-        p = subprocess.Popen("curl -O %s" % image_source, stdout=subprocess.PIPE, shell=True)
-        (output, err) = p.communicate()
+        exists = os.path.isfile(image_name)
+        if not exists:
+            p = subprocess.Popen("curl -O %s" % image_source, stdout=subprocess.PIPE, shell=True)
+            (output, err) = p.communicate()
+        else:
+            err = None
+
         if err == None:
 
             (res, metric_id) = resource_exists('image', resource_name)
